feat: Add draft survey questions
This is v1.0 of the survey. There are known bugs and it needs a redesign
already. Need to work out how to order the responses correctly and also
need to add user id to the responses so there can be multiple users. But
it's a start. Doing this commit before the redesign.