feat: Be able to select extra input type
When creating an option there can be options with extra input, e.g.
"other". For these types it must be specified which type of input such
as text or number. Instead of requiring the user to know the exact type
available and the exact text string for it to work, it's now presented
as a dropdown list.