fix: Survey list doesn't refresh on new survey
It is required to rotate or use another method to resume the activiy.
This fix pushes new surveys. Unfortunately for removing all surveys a
push doesn't come down as the entire directory is gone. But the users
won't get very far if they wanted to take the survy home for example.
That one is lower priority.